pyproject.toml
[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[project]
name = "e201-test-platform"
version = "0.1.0"
description = "Testing platform to test encoders with E201 interface"
readme = "README.md"
authors = [
    { name = "ropic", email = "nejc.ropic@rls.si" }
]
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.2",
    "pyqt5>=5.15.11",
    "pyserial>=3.5",
    "pyqtgraph"
]


[tool.uv]
native-tls = true
required-environments = [
    "sys_platform == 'win32' and platform_machine == 'AMD64'"
]

[project.scripts]
e201 = "e201_test_platform.main:main"